Added files to handle getting process information.
This patch provides the code to get process information for arbitrary
processes on various UNIX and UNIX-like systems.  For the moment, only
the GNU/Linux operating system is actually supported, and this code
has not yet been wired into Application.vala for use.

I don't have any other UNIX or UNIX-like systems to implement this on,
so implementations for FreeBSD, NetBSD, OpenBSD, OpenSolaris, and your
favorite UNIX or UNIX-like system near you are appreciated.