Enabled Linux-specific PPID/name retrieval code.
When starting up, we have to sometimes look for processes that we did
not expect to have to look for and get information on their process
ID, parent process ID, and program name.

The code that permits this to happen is enabled now, and used by
Application.vala.  This has the unfortunate effect of increasing our
dependency set over and above POSIX a little bit, and this is not good
because we lose portability.  As soon as possible, implementations to
do what we're doing for Linux need to be done for other UNIX-like
systems that AllTray can reasonably run on (e.g., systems that GNOME
runs on), such as FreeBSD, NetBSD, OpenBSD, OpenSolaris...

MonoDevelop can now be run under AllTray.  Unfortunately, it doesn't
seem that way for OpenOffice.org, though there may be a way to handle
that situation (where the shell script dies before the program
actually opens any windows, is what it looks like).