Subprocess support.
This enables AllTray to support processes spawned from a shell script
(in other words, we can attach to grandchildren now, instead of just
our children).

This doesn't work for OpenOffice.org, because while it is our
grandchild when it starts, the child dies before the grandchild's
window appears, and so OOo winds up with a ppid of 1, which defeats us
trying to attach to it.  The workaround is to use attach mode to
attach to it once it's already running.