Added support for displaying a different pointer over the CTT window.
For now, hard-coded to use the "default" theme, whatever that is for
any given individual's system.  Later, need to support the ability to
actually discover the cursor theme to use from the underlying WM or
DE, but since that would be an exercise in supporting everything under
the sun explicitly, it is not going to be done right now.